Upgrade to Delphi XE2 with Unicode and 64bit.
We did a decent job supporting Unicode before, but it's now better still, and the
implementation is a lot simpler. A number of custom interfaces to the Wide API,
and WideString versions of Delphi functions have been removed.

Some 64bit updates have also been done along the way, although there is still an
issue with dxgettext not working on that platform.

There is an issue with the dxgettext GnuGetText.pas unit, which I've replaced with
the newest version from their source control, but it still doesn't compile on XE2.
A couple of changes are required, which I'll commit separately to track them.