Fixed the application manifest.
The manifest is used to enable theme support on XP. It seems no
longer required on Windows 7 (Vista untested).

The old manifest did not specify 64bit as a target, caused the 64bit
shell extension to not work (maybe because of this), and was seemingly
not put in the right place for the activiation context to find it (at
resource id 2).